+++ package/Runtime/Scripts/DomeTools.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[HelpURL(DocumentationURL)]
+public class DomeTools : MonoBehaviour
+{
+    public const string DocumentationURL = "https://github.com/prefrontalcortex/DomeTools/blob/main/package/README.md";
+}